requirements.txt
Flask~=3.0.0
kinde-python-sdk==1.2.6
Flask-Session~=0.5.0
Werkzeug~=3.0.0
python-dotenv
psycopg2-binary
resend
pytailwindcss
Flask-Assets